muqss: Port "sched: Make nr_iowait() return 32-bit value"

+++ kernel/sched/MuQSS.c
@@ -3210,9 +3210,9 @@ unsigned long nr_iowait_cpu(int cpu)
  * Task CPU affinities can make all that even more 'interesting'.
  */
 
-unsigned long nr_iowait(void)
+unsigned int nr_iowait(void)
 {
-	unsigned long cpu, sum = 0;
+	unsigned int cpu, sum = 0;
 
 	for_each_possible_cpu(cpu)
 		sum += nr_iowait_cpu(cpu);
